database

+++ TransactionMicroservice/Models/Transfers.cs
@@ -15,7 +15,7 @@ namespace TransactionMicroservice.Models
         public int DestinationAccountId { get; set; }
 
         [Required]
-        [Range(1, int.MaxValue, ErrorMessage = "Amount can not be negative!")]
-        public int Amount { get; set; }
+        [Range(1, double.MaxValue, ErrorMessage = "Amount can not be negative!")]
+        public double Amount { get; set; }
     }
 }
